initial navbar dropdown and styling
event listeners and basic styling for the navbar dropdown, minor change to navbar button. some links in dropdown are not implemented yet. visuals may change in the future, but I think this is a decent placeholder for now. a couple of pages have some minor styling issues but this doesn't impact functionality, so this can be dealt with later 👍